refactor(be): move class and interfaces to appropriate dir
