+++ apps/backend/apps/client/src/submission/mock/submissionResult.mock.ts
@@ -1,5 +1,5 @@
 import { ResultStatus, type SubmissionResult } from '@prisma/client'
-import type { JudgerResponse } from '../dto/judger-response.dto'
+import type { JudgerResponse } from '../class/judger-response.dto'
 
 export const submissionResults: SubmissionResult[] = [
   {
