+++ apps/backend/apps/client/src/submission/submission.controller.ts
@@ -24,7 +24,7 @@ import {
   IDValidationPipe,
   RequiredIntPipe
 } from '@libs/pipe'
-import { CreateSubmissionDto } from './dto/create-submission.dto'
+import { CreateSubmissionDto } from './class/create-submission.dto'
 import { SubmissionService } from './submission.service'
 
 @Controller('submission')
